third_party/WebKit/LayoutTests/dom/document_type/previous-element-sibling.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<script>

description('This tests the NonDocumentTypeChildNode interface previousElementSibling property on a DocumentType node.');

var doctype = document.doctype;
shouldBeUndefined('doctype.previousElementSibling');

</script>
