third_party/WebKit/LayoutTests/dom/document_type/previous-element-sibling-expected.txt

This tests the NonDocumentTypeChildNode interface previousElementSibling property on a DocumentType node.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS doctype.previousElementSibling is undefined.
PASS successfullyParsed is true

TEST COMPLETE
